Replace StopListening with Exit in IUdpClient
Refactored the `IUdpClient` interface by removing the `StopListening` method and introducing the `Exit` method. Updated the `UdpClientWrapper` class to implement the new `Exit` method, replacing the functionality of `StopListening`.

Modified the `NetSdrClient` class to use `_udpClient.Exit()` instead of `_udpClient.StopListening()` to align with the updated interface.

Updated unit tests in `UdpClientWrapperTests` to validate the behavior of the new `Exit` method, replacing references to `StopListening` in assertions and test descriptions.